pyproject.toml

[tool.poetry]
name = "carte-grise-ocr"
version = "0.1.0"
description = "FastAPI backend for car plate OCR extraction"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.6.1"
sqlalchemy = "^2.0.25.post1"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
celery = "^5.3.6"
redis = "^5.0.1"
opencv-python = "^4.9.0.80"
pillow = "^10.2.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
mistralai = "^0.0.10"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.25.2"
pytest-mock = "^3.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
